ErrorControl: Fix non-portable cast.
This should enable 64-bit-to-64-bit multiplayer matches. I also reviewed ErrorControl, and I didn't see any other casts as found here. There are usages of long in the encoded messages, which will affect 64-bit-to-32-bit, but that is a problem for another day.

ErrorControl is a bit redundant in a couple aspects, and could simply be removed, with some testing. It was likely used in situations that are no longer relevant.

+++ include/OERRCTRL.h
@@ -118,7 +118,7 @@ public:
 	char	get_ec_player_id( unsigned long dpPlayerId );
 
 	int	send(char ecPlayerId, void *dataPtr, unsigned long dataLen);
-	char *receive(char *sendEcPlayerId, unsigned long *dataLen);
+	char*	receive(char *sendEcPlayerId, uint32_t *dataLen);
 	void	de_recv_queue();			// get the content from recv_queue[recv_head] before de_recv_queue
 
 	int	is_player_valid(char ecPlayerId);
